refactor: 사용하지 않는 메소드 삭제 및 메소드명 변경
